insight-bot/insight-bot/requirements.txt
praw==7.7.1
anthropic==0.25.0
pymongo==4.6.1
python-dotenv==1.0.1
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2
# Don't forget to run: python -m spacy download en_core_web_sm